Disable InOrder when running pipelines in test mode
Summary:
InOrder uses a thread-local to make sure you don't have an InOrder nested within
another InOrder, which works normally because each pipeline is run in a separate
process. In test mode, though, everything is run in one thread, so this
reentrancy check incorrectly throws an exception. Since InOrder has no effect in
test mode anyway, I fixed this by just disabling InOrder for tests.

Test Plan:
This was breaking a new test in webapp for some code that had been running
correctly in production mode for a while. With this commit, the test works.

Reviewers: mattfaus

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D7815