Remove an incorrect assertion in pipeline test mode
Summary:
The implementation of evaluate_test had an assertion that all outputs of a
pipeline are always consumed by the calling pipeline, but as far as I can tell,
this isn't enforced for regular production mode, so maybe the test assertion was
left in by accident. It's breaking a test on a pipeline that has been running
fine for a while, so I'm removing the test sanity check.

Test Plan: Run summarize_fast_ete_test in webapp, which ran into this bug.

Reviewers: mattfaus

Reviewed By: mattfaus

Subscribers: benkraft, chris

Differential Revision: http://phabricator.khanacademy.org/D10679