Change /_ah/pipeline base_path to just /_pipeline
Summary:
See [1] for context. This is precisely the patch at [2], but I
cherry-picked it instead of merging it because the remote is more
up-to-date with upstream/master than we are, and I don't want other
changes in pipelines to break our code.

Depends on D29175.

[1]: https://github.com/GoogleCloudPlatform/appengine-pipelines/issues/57
[2]: https://github.com/someone1/appengine-pipelines/commit/3345cdd30fc3bf45e6f61dfaf6634cc2d810c665

Test Plan:
None. I tried to run the tests, but they gave me a bunch of PYTHONPATH
errors, so I gave up. I can try to get them to run tomorrow. But, other
than the test code, this is a one-line change that //should// be
forward-compatible given D29175.

I'm assuming that none of the changes to the demo app will impact
production.

Reviewers: chris, csilvers

Reviewed By: csilvers

Subscribers: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D29178