Add a note about master_with_khan_patches
Summary:
I just spent a good deal of time trying to debug a Heisenbug with the
substate for this repo. It turns out that although `origin` points to
our Khan/appengine-pipelines fork, the webapp substate for this repo is
//not// `origin/master`, but rather `origin/master_with_khan_patches`.
This is pretty darn unexpected, so for the time being I'm just leaving a
note that no one will read.

Arguably, we should just delete the master branch and move
master_with_khan_patches there—this is our fork, after all—but perhaps
Khan/appengine-pipelines is popular enough that others rely upon it.

Test Plan: None.

Reviewers: chris, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D29187