Route in test mode based on the server environment
Summary:

After upgrading the appengine SDK to 1.9.54, several of our pipeline
tests broke.  This is a known issue with an outstanding pull request:
https://github.com/GoogleCloudPlatform/appengine-pipelines/pull/78

However, the fix from the pull request, which I applied in
https://phabricator.khanacademy.org/rGAPPI1c757b3d21eb9fe5bbf4a9c503a6292a2de6a15a
doesn't appear to be the right fix for us, as setting `_TEST_MODE` has
other side effects that we don't want.

Instead, we want to *route* as if we're in _TEST_MODE-- i.e. set
_TEST_MODE based on the server software in that function only-- but not
set _TEST_MODE everywhere.

Test plan:
- substate into webapp
- `make check`

Auditors: csilvers